mlops-app/requirements.txt
dagshub==0.3.34
datasets==2.20.0
flask==3.0.3
hydra-core==1.3.2
mlflow==2.14.1
torch==2.3.1
transformers==4.41.2